Porting the project to CMake and implementing a lot of optimisations
