--- AutoSprint.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30621.155
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "AutoSprint", "AutoSprint\AutoSprint.vcxproj", "{E9333B08-B224-4849-AE9D-E9845DA5C3E6}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E9333B08-B224-4849-AE9D-E9845DA5C3E6}.Debug|x64.ActiveCfg = Debug|x64
-		{E9333B08-B224-4849-AE9D-E9845DA5C3E6}.Debug|x64.Build.0 = Debug|x64
-		{E9333B08-B224-4849-AE9D-E9845DA5C3E6}.Debug|x86.ActiveCfg = Debug|Win32
-		{E9333B08-B224-4849-AE9D-E9845DA5C3E6}.Debug|x86.Build.0 = Debug|Win32
-		{E9333B08-B224-4849-AE9D-E9845DA5C3E6}.Release|x64.ActiveCfg = Release|x64
-		{E9333B08-B224-4849-AE9D-E9845DA5C3E6}.Release|x64.Build.0 = Release|x64
-		{E9333B08-B224-4849-AE9D-E9845DA5C3E6}.Release|x86.ActiveCfg = Release|Win32
-		{E9333B08-B224-4849-AE9D-E9845DA5C3E6}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {1B18EEA6-E00B-4D44-9CEE-BB7FA3EF671B}
-	EndGlobalSection
-EndGlobal
