--- AutoSprint/framework.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#define WIN32_LEAN_AND_MEAN
-
-#include <windows.h>
