--- AutoSprint/pch.h
@@ -1,6 +1,0 @@
-#ifndef PCH_H
-#define PCH_H
-
-#include "framework.h"
-
-#endif //PCH_H
